main: Subscribe to changes in configs
In atom, we need to subscribe to the value of configurations to be able to
use them on the fly. Earlier although the configuration was there, atom had
to be restarted for the value to be used. Now, an event trigger is used
instead to change the executable value on the fly.