main.coffee: Correct parsing regex
This addition was intended to take care of trailing newlines it seems.
Despite us knowing that this plugin actually worked for coala 0.4 and
probably up until 0.6 or 0.8 it is currently broken.

`apm test` shows this as well and fails before this commit but passes
after locally.

The regex was introduced in 52a8d77ffa784d037f8bcebfd940bfb7b35920ea.